[front] Move component test cases to __tests__ dir
