+++ frontend/src/components/__tests__/Question/Form/ActionBox.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 import React from 'react';
 import { fireEvent, render } from 'react-testing-library';
-import ActionBox from './ActionBox';
+import ActionBox from '../../../Question/Form/ActionBox';
 import Router from 'next/router';
 
 jest.mock('next/router');
